Remove duplicate exchange rate
